Refactor x86 module
