+++ src/sys/idt.rs
@@ -2,8 +2,8 @@ use crate::{api, hang, sys};
 use crate::api::process::ExitCode;
 use crate::sys::pic;
 use crate::sys::process::Registers;
-use crate::sys::x86::interrupts;
-use crate::sys::x86::registers::Cr2;
+use crate::sys::x86::int;
+use crate::sys::x86::reg::Cr2;
 
 use core::arch::{asm, naked_asm};
 use lazy_static::lazy_static;
@@ -241,7 +241,7 @@ extern "sysv64" fn syscall_handler(
 }
 
 pub fn set_irq_handler(irq: u8, handler: fn()) {
-    interrupts::without_interrupts(|| {
+    int::without_interrupts(|| {
         let mut handlers = IRQ_HANDLERS.lock();
         handlers[irq as usize] = handler;
