+++ src/sys/log.rs
@@ -1,4 +1,4 @@
-use crate::sys::x86::interrupts;
+use crate::sys::x86::int;
 
 use alloc::string::String;
 use core::fmt;
@@ -51,13 +51,13 @@ impl core::fmt::Write for LogBuffer {
 
 #[doc(hidden)]
 pub fn write_fmt(args: fmt::Arguments) {
-    interrupts::without_interrupts(||
+    int::without_interrupts(||
         LOG.lock().write_fmt(args).expect("Could not write log")
     )
 }
 
 pub fn read() -> String {
-    interrupts::without_interrupts(|| {
+    int::without_interrupts(|| {
         let log = LOG.lock();
         String::from_utf8_lossy(log.buf()).into_owned()
     })
