+++ src/sys/vga/font.rs
@@ -2,7 +2,7 @@ use super::writer::WRITER;
 
 use crate::api::font::Font;
 use crate::api::fs::{FileIO, IO};
-use crate::sys::x86::interrupts;
+use crate::sys::x86::int;
 
 use core::convert::TryFrom;
 use spin::Mutex;
@@ -44,7 +44,7 @@ impl FileIO for VgaFont {
 }
 
 fn write_font(font: &Font) {
-    interrupts::without_interrupts(||
+    int::without_interrupts(||
         WRITER.lock().set_font(font)
     )
 }
